+++ src/main/java/insane96mcp/shieldsplus/setup/client/Client.java
@@ -10,6 +10,9 @@ import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import net.neoforged.neoforge.event.BuildCreativeModeTabContentsEvent;
 import net.neoforged.neoforge.registries.DeferredHolder;
 
+import static insane96mcp.insanelib.util.CreativeTabsUtils.addAfter;
+import static insane96mcp.insanelib.util.CreativeTabsUtils.addBefore;
+
 public class Client {
 
     public static void setup(final FMLClientSetupEvent event) {
@@ -37,13 +40,5 @@ public class Client {
             addAfter(event, Items.SHIELD, SPItems.GOLDEN_SHIELD.get());
             event.remove(new ItemStack(Items.SHIELD), CreativeModeTab.TabVisibility.PARENT_AND_SEARCH_TABS);
         }
-    }
-
-    private static void addBefore(BuildCreativeModeTabContentsEvent event, Item before, Item itemToAdd) {
-        event.insertBefore(new ItemStack(before), new ItemStack(itemToAdd), CreativeModeTab.TabVisibility.PARENT_AND_SEARCH_TABS);
-    }
-
-    private static void addAfter(BuildCreativeModeTabContentsEvent event, Item after, Item itemToAdd) {
-        event.insertAfter(new ItemStack(after), new ItemStack(itemToAdd), CreativeModeTab.TabVisibility.PARENT_AND_SEARCH_TABS);
     }
 }